# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: AI SkillFit API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /candidates:
    post:
      operationId: registerCandidate
      tags: [candidates]
      summary: Register a new candidate
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RegisterCandidateBody"
      responses:
        "201":
          description: Registered candidate
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Candidate"
        "409":
          description: Duplicate candidate
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
    get:
      operationId: listCandidates
      tags: [candidates]
      summary: List all candidates (admin)
      parameters:
        - name: district
          in: query
          schema:
            type: string
        - name: trade
          in: query
          schema:
            type: string
        - name: language
          in: query
          schema:
            type: string
        - name: classification
          in: query
          schema:
            type: string
        - name: search
          in: query
          schema:
            type: string
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: limit
          in: query
          schema:
            type: integer
            default: 20
      responses:
        "200":
          description: Paginated candidates list
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CandidateListResponse"
  /candidates/{id}:
    get:
      operationId: getCandidate
      tags: [candidates]
      summary: Get candidate detail with interview data
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Candidate detail
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CandidateDetail"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
...
components:
  schemas:
...
    ApiError:
      type: object
      properties:
        error:
          type: string
      required:
        - error
    RegisterCandidateBody:
      type: object
      properties:
        name:
          type: string
        phone:
          type: string
        district:
          type: string
        trade:
          type: string
        language:
          type: string
          enum: [kn, hi, en]
        deviceFingerprint:
          type: string
      required:
        - name
        - phone
        - district
        - trade
        - language
    Candidate:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        phone:
          type: string
        district:
          type: string
        trade:
          type: string
        language:
          type: string
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - name
        - phone
        - district
        - trade
        - language
        - createdAt
    CandidateListResponse:
      type: object
      properties:
        candidates:
          type: array
          items:
            $ref: "#/components/schemas/CandidateWithClassification"
        total:
          type: integer
        page:
          type: integer
        limit:
          type: integer
      required:
        - candidates
        - total
        - page
        - limit
    CandidateWithClassification:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        phone:
          type: string
        district:
          type: string
        trade:
          type: string
        language:
          type: string
        createdAt:
          type: string
          format: date-time
        classification:
          type: string
          nullable: true
        avgScore:
          type: number
          nullable: true
        interviewId:
          type: integer
          nullable: true
        interviewStatus:
          type: string
          nullable: true
      required:
        - id
        - name
        - phone
        - district
        - trade
        - language
        - createdAt
    CandidateDetail:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        phone:
          type: string
        district:
          type: string
        trade:
          type: string
        language:
          type: string
        createdAt:
          type: string
          format: date-time
        interview:
          $ref: "#/components/schemas/InterviewDetail"
          nullable: true
        classification:
          $ref: "#/components/schemas/Classification"
          nullable: true
        integrityCheck:
          $ref: "#/components/schemas/IntegrityCheck"
          nullable: true
        officerActions:
          type: array
          items:
            $ref: "#/components/schemas/OfficerAction"
      required:
        - id
        - name
        - phone
        - district
        - trade
        - language
        - createdAt
        - officerActions
...
    InterviewDetail:
      type: object
      properties:
        id:
          type: integer
        candidateId:
          type: integer
        status:
          type: string
        questionSnapshot:
          type: string
          nullable: true
        createdAt:
          type: string
          format: date-time
        completedAt:
          type: string
          format: date-time
          nullable: true
        responses:
          type: array
          items:
            $ref: "#/components/schemas/Response"
      required:
        - id
        - candidateId
        - status
        - createdAt
        - responses
...
    Response:
      type: object
      properties:
        id:
          type: integer
        interviewId:
          type: integer
        questionId:
          type: string
        questionText:
          type: string
        transcript:
          type: string
        videoUrl:
          type: string
          nullable: true
        relevanceScore:
          type: number
          nullable: true
        clarityScore:
          type: number
          nullable: true
        confidenceScore:
          type: number
          nullable: true
        geminiReasoning:
          type: string
          nullable: true
        languageDetected:
          type: string
          nullable: true
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - interviewId
        - questionId
        - questionText
        - transcript
        - createdAt
...
    Classification:
      type: object
      properties:
        id:
          type: integer
        interviewId:
          type: integer
        category:
          type: string
          enum: [job_ready, requires_training, manual_verification, poor_quality, suspected_duplicate]
        avgScore:
          type: number
        reasoning:
          type: string
        reviewedBy:
          type: string
          nullable: true
        reviewedAt:
          type: string
          format: date-time
          nullable: true
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - interviewId
        - category
        - avgScore
        - reasoning
        - createdAt
    IntegrityCheck:
      type: object
      properties:
        id:
          type: integer
        interviewId:
          type: integer
        facePresentPct:
          type: number
          nullable: true
        livenessPass:
          type: boolean
          nullable: true
        duplicateCandidateId:
          type: integer
          nullable: true
        flags:
          type: array
          items:
            type: string
      required:
        - id
        - interviewId
        - flags
...
    OfficerAction:
      type: object
      properties:
        id:
          type: integer
        candidateId:
          type: integer
        interviewId:
          type: integer
          nullable: true
        action:
          type: string
        officerUsername:
          type: string
        notes:
          type: string
          nullable: true
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - candidateId
        - action
        - officerUsername
        - createdAt
...